+++ src/array/codec/array_to_bytes/bytes/bytes_partial_decoder.rs
@@ -127,6 +127,13 @@ impl AsyncArrayPartialDecoderTraits for AsyncBytesPartialDecoder<'_> {
         let mut bytes = Vec::with_capacity(decoded_regions.len());
         let chunk_shape = self.decoded_representation.shape_u64();
         for array_subset in decoded_regions {
+            if array_subset.dimensionality() != self.decoded_representation.dimensionality() {
+                return Err(CodecError::InvalidArraySubsetDimensionalityError(
+                    array_subset.clone(),
+                    self.decoded_representation.dimensionality(),
+                ));
+            }
+
             // Get byte ranges
             let byte_ranges = array_subset
                 .byte_ranges(&chunk_shape, self.decoded_representation.element_size())