+++ src/array/codec/array_to_bytes/sharding/sharding_partial_decoder.rs
@@ -131,6 +131,15 @@ impl ArrayPartialDecoderTraits for ShardingPartialDecoder<'_> {
         array_subsets: &[ArraySubset],
         options: &PartialDecodeOptions,
     ) -> Result<Vec<Vec<u8>>, CodecError> {
+        for array_subset in array_subsets {
+            if array_subset.dimensionality() != self.decoded_representation.dimensionality() {
+                return Err(CodecError::InvalidArraySubsetDimensionalityError(
+                    array_subset.clone(),
+                    self.decoded_representation.dimensionality(),
+                ));
+            }
+        }
+
         let Some(shard_index) = &self.shard_index else {
             return Ok(array_subsets
                 .iter()